added update balance functionality

+++ src/models/schema.js
@@ -79,7 +79,7 @@ export const schema = {
                 "amount": {
                     "name": "amount",
                     "isArray": false,
-                    "type": "Int",
+                    "type": "Float",
                     "isRequired": false,
                     "attributes": []
                 },
@@ -100,7 +100,7 @@ export const schema = {
                 "balance": {
                     "name": "balance",
                     "isArray": false,
-                    "type": "Int",
+                    "type": "Float",
                     "isRequired": false,
                     "attributes": []
                 },
